Massive Changes

+++ src/Renderer/Vertices/_Vertex.h
@@ -16,9 +16,9 @@ struct Vertex {
     int16_t posX, posY, posZ;
     char texU, texV; // Local UV tiling factor (repeat count)
     char direction;
-    uint8_t layerIndex; // Texture Layer Index (0-255)
+    uint16_t layerIndex; // Texture Layer Index (0-65535)
 
-    Vertex(int16_t _posX, int16_t _posY, int16_t _posZ, char _texU, char _texV, char _direction, uint8_t _layerIndex)
+    Vertex(int16_t _posX, int16_t _posY, int16_t _posZ, char _texU, char _texV, char _direction, uint16_t _layerIndex)
         : posX(_posX), posY(_posY), posZ(_posZ), texU(_texU), texV(_texV), direction(_direction), layerIndex(_layerIndex) {}
 };
 #pragma pack(pop)
